rename search places fn
