+++ libs/expo/betterangels/src/lib/services/PlaceAutocomplete/constants.google.maps.ts
@@ -1,0 +1,10 @@
+import { TPlaceLatLng } from './types';
+
+// geocode for approx center of LA COUNTY
+export const LA_COUNTY_LAT = 34.04499;
+export const LA_COUNTY_LNG = -118.251601;
+
+export const LA_CENTER_LAT_LNG: TPlaceLatLng = {
+  lat: LA_COUNTY_LAT,
+  lng: LA_COUNTY_LNG,
+};
