+++ libs/expo/betterangels/src/lib/services/PlaceAutocomplete/index.ts
@@ -1,3 +1,4 @@
+export * from './constants.google.maps';
+export { getPlaceAutocomplete } from './getPlaceAutocomplete';
 export { getPlaceDetailsById } from './getPlaceDetailsById';
-export { searchPlacesLA } from './searchPlacesLA';
 export * from './types';
